+++ src/main/java/org/ikasan/studio/ui/Styling.java
@@ -1,0 +1,11 @@
+package org.ikasan.studio.ui;
+
+import java.awt.*;
+
+public class Styling {
+    //    public static final Color IKASAN_GREY = Gray._231;
+        public static final Color IKASAN_GREY = new Color(238,238,238);
+    public static final Color IKASAN_ORANGE = new Color(241, 90, 35);
+    public static final Color IKASAN_RED = Color.RED;
+    public static final Color IKASAN_BLACK = Color.BLACK;
+}