+++ CHANGELOG.md
@@ -31,6 +31,7 @@ UX
 - Fix bug in startup meta pack
 - For new Maven generated projects, the app name and package name are pre-populated from the maven values.
 - Swap back to palette panel when property update is completed.
+- Add UI helper for any cron based data entry
 
 metapackVersion